pkg/validate/validate: Move SSH validation into pkg/validate
This logic is independent of our asset package.
